backend/requirements.txt
flask
flask-cors
opencv-python-headless
pillow
numpy
scikit-image
dahuffman
imagecodecs
matplotlib
ultralytics
pytest
